fix(router): measure via-segment demote against committed net_routes, not stale grid list
Addresses PR #3654 review (judge changes-requested):

1. Rebased onto origin/main; resolved the route_all_negotiated
   finalization conflict so #3628's _unwind_correction_pass_net and this
   PR's _demote_via_segment_violation_nets both coexist (distinct methods;
   the via-segment demote stays AFTER #3628's correction-pass unwind in
   the seg-seg -> pad-clearance -> via-segment demote sequence).

2. Board-02 regression root cause + fix (the crux): the +9-via / +25 mm
   drift was NOT #3020 oscillation -- it was a STALE-UNIVERSE read.
   worst_via_segment_deficit measured the foreign-segment universe from
   the grid's own self.routes list, but the best-iteration restore
   (#2540/#2803) rewinds self.routes via the *_usage calls only while
   re-syncing the authoritative net_routes.  After a restore the grid
   list still held a worse iteration's geometry, so the backstop saw
   false-positive shorts on NODE_B/NODE_C, demoted two clean nets to
   6/8, and forced the escalation ladder to a 4-layer / via-heavy
   solution.  The new foreign_routes= parameter makes the demote measure
   against the committed net_routes -- bit-for-bit with the in-loop #3020
   hook, which already builds its universe from net_routes.  Board 02 is
   back to baseline (22 routes, 24 vias, 328.16 mm, 8/8, no seg-seg
   shorts) and the softstart geometry tests still pass, so a genuine
   residual short is still demoted.

3. Removed orphaned tests/test_softstart_segment_via_repair.py (it
   exercised the deleted _repair_segment_via_clearance /
   _point_segment_distance workaround).  Its artifact-level invariants
   (committed-board barrel-clearance scan, 45-alignment, fleet-wide uuid
   uniqueness) were NOT covered elsewhere, so they are preserved in the
   new tests/test_softstart_artifact_barrel_clearance.py, which depends
   only on the committed .kicad_pcb files and is-45-aligned helper.
